@@ -10,7 +10,7 @@ import com.networknt.email.EmailSender;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.mail.MessagingException;
+import jakarta.mail.*;
 import java.io.File;
 import java.util.Map;
 import java.util.Set;
